% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_char_testphrases}
\alias{data_char_testphrases}
\title{Sample short documents for testing}
\format{An object of class \code{character} of length 10.}
\usage{
data_char_testphrases
}
\description{
Some sample short documents in plain text format for testing
  with \code{\link{liwcalike}}.
}
\examples{
liwcalike(data_char_testphrases)
}
\keyword{datasets}
