% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_AFINN}
\alias{data_dictionary_AFINN}
\title{Affective Norms for English Words (AFINN)}
\format{An object of class \code{dictionary2} of length 2.}
\source{
\url{http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010}
}
\usage{
data_dictionary_AFINN
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing
the Affective Norms for English Words (AFINN), a publicly available list of 
English words rated for valence with values between -5 (negative) and
+5 (positive). AFINN-111, the latest version, contains 2,477 words and phrases. 
Note that the version implemented in \pkg{quanteda.dictionaries} uses a binary 
classification into 878 positive (valence > 0) and 1,599 negative (valence < 0) 
features.
}
\references{
Nielsen, Finn Å. (2011). "\href{https://arxiv.org/abs/1103.2903}{A new ANEW: 
  Evaluation of a Word List for Sentiment Analysis in Microblogs.}" 
  In \emph{Proceedings of the ESWC2011 Workshop on 'Making Sense of Microposts': 
  Big Things Come in Small Packages}: 93-98.
}
\keyword{data}
