% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_geninqposneg}
\alias{data_dictionary_geninqposneg}
\title{augmented General Inquirer \emph{Positiv} and \emph{Negativ} dictionary}
\format{An object of class \code{dictionary2} of length 2.}
\source{
\url{http://www.wjh.harvard.edu/~inquirer/spreadsheet_guide.htm}
}
\usage{
data_dictionary_geninqposneg
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing the
\emph{Positiv} and \emph{Negativ} dictionary entries from the augmented
General Inquirer.  These are new valence categories described at
\url{http://www.wjh.harvard.edu/~inquirer/homecat.htm} but also including the
terms from the \href{http://www.wjh.harvard.edu/~inquirer/Yes.html}{yes} and
\href{http://www.wjh.harvard.edu/~inquirer/No.html}{no} dictionary entries.
}
\references{
Stone, Philip J., Dexter C. Dunphy, and Marshall S. Smith. 1966.
  "The general inquirer: A computer approach to content analysis." Cambridge,
  MA: MIT Press.
}
\keyword{data}
