% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_LoughranMcDonald}
\alias{data_dictionary_LoughranMcDonald}
\title{Loughran and McDonald Sentiment Word Lists}
\format{An object of class \code{dictionary2} of length 9.}
\source{
\url{http://sraf.nd.edu/textual-analysis/resources/}
}
\usage{
data_dictionary_LoughranMcDonald
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing
the 2014 version of the Loughran and McDonald Sentiment Word Lists. The
categories are "negative" (2355 features), "positive" (354), "uncertainty" (297), "litigious" (903),
"constraining" (184), "superfluous" (56), "interesting" (68), "modal words strong" (68)
and "modal words weak" (0).
}
\references{
Loughran, T. & McDonald, B. (2011). \href{https://doi.org/10.1111/j.1540-6261.2010.01625.x}{When 
  is a Liability not a Liability? Textual Analysis, Dictionaries, and 10-Ks}. 
  \emph{Journal of Finance}, 66(1), 35--65.
}
\keyword{data}
