% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_NRC}
\alias{data_dictionary_NRC}
\title{NRC Word-Emotion Association Lexicon}
\format{An object of class \code{dictionary2} of length 10.}
\usage{
data_dictionary_NRC
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing Mohammad and
Charron's (2010, 2013)  NRC Word-Emotion Association Lexicon (aka NRC Emotion
Lexicon aka EmoLex): association of words with eight emotions (anger, fear,
anticipation, trust, surprise, sadness, joy, and disgust) and two sentiments
(negative and positive) manually annotated on Amazon's Mechanical Turk.
Available in 40 different languages.
}
\references{
Saif Mohammad and Peter Turney (2013). "Crowdsourcing a
  Word-Emotion Association Lexicon."  \emph{Computational Intelligence} 29
  (3), 436-465.

  Saif Mohammad and Peter Turney (2010). "Emotions Evoked by Common Words and
  Phrases: Using Mechanical Turk to Create an Emotion Lexicon." In
  \emph{Proceedings of the NAACL-HLT 2010 Workshop on Computational
  Approaches to Analysis and Generation of Emotion in Text}, June 2010, LA,
  California.

  NRC webpage: \url{http://saifmohammad.com/WebPages/AccessResource.htm}
}
\keyword{data}
