% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_Rauh}
\alias{data_dictionary_Rauh}
\title{Rauh's German Political Sentiment Dictionary}
\format{The dictionary has four keys.
\describe{
\item{\code{negative}}{19,750 terms indicating negative sentiment}
\item{\code{positive}}{17,330 terms indicating positive sentiment}
\item{\code{neg_positive}}{17,330 terms indicating a positive word preceded by a negation (used to convey negative sentiment)}
\item{\code{neg_negative}}{19,750 terms indicating a negative word preceded by a negation (used to convey positive sentiment)}
}}
\source{
\url{https://doi.org/10.7910/DVN/BKBXWD}
}
\usage{
data_dictionary_Rauh
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing 
the dictionaries provided in Rauh (forthcoming). The augmented German sentiment
dictionary is based on SentiWS and GermanPolarityClues. Rauh assesses its performance against human intuition in parliamentary speeches, party manifestos and media coverage.
Note that a combination of \link[quanteda]{tokens_replace} and \link[quanteda]{tokens_compound} is currently required to harmonize the five German bi-gram negations. The example below show how to conduct this transformation. Note that the process changes the the terms "nicht|nichts|kein|keine|keinen" to joint term and thus replaces some of the original features of the corpus.
}
\examples{
replace German negation terms and compound neg_negative and neg_positive
toks <- quanteda::tokens("nicht schlecht dieses wunderschöne Wetter")

# replace negations with "not".
toks1 <- quanteda::tokens_replace(toks, pattern = c("nicht", "nichts", "kein", "keine", "keinen"), rep("not", 5)) 

# compound tokens
toks2 <- quanteda::tokens_compound(toks1, data_dictionary_rauh, concatenator = " ")

# apply dictionary
toks2 \%>\% 
    quanteda::dfm() \%>\% 
    quanteda::dfm_lookup(data_dictionary_rauh) 
}
\references{
Rauh, C. (Forthcoming). "Validating a Sentiment Dictionary for German Political 
  Language: A Workbench Note." \emph{Journal of Information Technology & Politics}.
}
\keyword{data}
