% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_RID}
\alias{data_dictionary_RID}
\title{Martindale's Regressive Imagery Dictionary (RID)}
\format{The dictionary has three levels of nesting, and forms 43 keys in
  total and a total of 3,150 values.}
\usage{
data_dictionary_RID
}
\description{
The English Regressive Imagery Dictionary (RID) consists of about 3,150 words
and roots assigned to 29 categories of primary process cognition, 7
categories of secondary process cognition, and 7 categories of emotions,
designed by Martindale (1975,1990) to measure primordial vs. conceptual
thinking.
}
\references{
Martindale, C. (1975). \emph{Romantic progression: The psychology
of literary history.} Washington, D.C.: Hemisphere.

Martindale, C. (1990). \emph{The clockwork muse: The predictability of
artistic change.} New York: Basic Books.
}
\keyword{data}
