% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.dictionaries-package.r
\docType{package}
\name{quanteda.dictionaries}
\alias{quanteda.dictionaries}
\alias{quanteda.dictionaries-package}
\title{quanteda.dictionaries}
\description{
Provides text analysis dictionaries and additional functionality for their
use in text analysis frameworks, especially quanteda.
}
