% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readliwc.R
\name{readliwc}
\alias{readliwc}
\title{Create a dictionary from a LIWC dictionary poster pdf}
\usage{
readliwc(file)
}
\arguments{
\item{file}{the filename of the LIWC dictionary poster pdf to be read}
}
\value{
a \pkg{quanteda} \link[quanteda]{dictionary}
}
\description{
Creates a \pkg{quanteda} dictionary from a LIWC dictionary output from the
LIWC software as a "dictionary poster" pdf.  Currently tested with LIWC 2015
v1.5.0, available for purchase from \url{http://http://liwc.wpengine.com/}.
The poster files can be output using Dictionary -> Export Internal
Dictionaries from the LIWC2015 application menu.
}
\examples{
\dontrun{
data_dictionary_liwc2015eng <- readliwc("~/Desktop/LIWC2015 dictionary poster.pdf")
names(data_dictionary_liwc2015eng)
data_dictionary_liwc2015eng["Assent"]
data_dictionary_liwc2015eng["Netspeak"]
}
}
