% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_HuLiu}
\alias{data_dictionary_HuLiu}
\title{positive and negative words from Hu and Liu (2004)}
\format{An object of class \code{dictionary2} of length 2.}
\source{
\url{http://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html}
}
\usage{
data_dictionary_HuLiu
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing the positive
and negative words from Hu and Liu (2004, 2005).
}
\references{
Minqing Hu and Bing Liu. "Mining and Summarizing Customer
  Reviews." Proceedings of the ACM SIGKDD International Conference on
  Knowledge Discovery and Data Mining (KDD-2004), Aug 22-25, 2004, Seattle,
  Washington, USA,

  Bing Liu, Minqing Hu and Junsheng Cheng. "Opinion Observer: Analyzing and
  Comparing Opinions on the Web." Proceedings of the 14th International World
  Wide Web conference (WWW-2005), May 10-14, 2005, Chiba, Japan. Young, L.
  and Soroka, S. 2012. \emph{Lexicoder Sentiment Dictionary}. Available at
  \url{http://lexicoder.com}.
}
\keyword{data}
