% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_MFD}
\alias{data_dictionary_MFD}
\title{Moral Foundations Dictionary}
\format{An object of class \code{dictionary2} of length 10.}
\source{
http://www.jeremyfrimer.com/research-downloads.html; a previous
  version is available at \url{http://moralfoundations.org/othermaterials}
}
\usage{
data_dictionary_MFD
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing
the Moral Foundations Dictionary, a publicly available dictionaries with
information on the proportions of virtue and vice words for each foundation.
The categories are harm (vice/virtue), fairness (vice/virtue), ingroup (vice/virtue),
authority (vice/virtue), purity (vice/virtue) and morality (general).

This version is 2.0 version of the dictionary,
  \href{http://www.jeremyfrimer.com/uploads/2/1/2/7/21278832/summary.pdf}{recommended}
  over the first version of the MDF by its authors.
}
\references{
Frimer, J. et. al. (2017).  Moral Foundations Dictionaries for
  Linguistic Analyses, 2.0. University of Winnipeg Manuscript.
  \url{http://www.jeremyfrimer.com/uploads/2/1/2/7/21278832/summary.pdf}.

  Haidt, J., Graham, J., & Nosek, B.A. (2009). \href{http://dx.doi.org/10.1037/a0015141}{Liberals 
  and Conservatives Rely on Different Sets of Moral Foundations}. 
  \emph{Journal of Personality and Social Inquiry}, 20(2--3), 110--119.

  Graham, J., & Haidt, J. (2016). \href{http://moralfoundations.org/othermaterials}{Moral
  Foundations Dictionary.}: \url{http://moralfoundations.org/othermaterials}.
}
\keyword{data}
