% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_sentiws}
\alias{data_dictionary_sentiws}
\title{SentimentWortschatz (SentiWS)}
\format{An object of class \code{dictionary2} of length 2.}
\source{
\url{http://wortschatz.uni-leipzig.de/en/download/}
}
\usage{
data_dictionary_sentiws
}
\description{
A \pkg{quanteda} \link[quanteda]{dictionary} object containing
SentimentWortschatz (SentiWS), a publicly available German-language resource
for sentiment analysis. The current version of SentiWS contains 1,650
positive and 1,818 negative words, which sum up to 15,649 positive and 15,632
negative word forms including their inflections. It not only contains
adjectives and adverbs explicitly expressing a sentiment, but also nouns and
verbs implicitly containing one. The original dictionary weights within the
interval of -1 to 1. Note that the version implemented in
\pkg{quanteda.dictionaries} uses a binary classification into positive
(weight > 0) and negative (weight < 0) features.
}
\references{
Remus, R., Quasthoff U., and Heyer, G. (2010).
  "\href{http://www.lrec-conf.org/proceedings/lrec2010/pdf/490_Paper.pdf}{SentiWS
  - a Publicly Available German-language Resource for Sentiment Analysis.}"
  In \emph{Proceedings of the 7th International Language Ressources and
  Evaluation (LREC'10)}: 1168-1171.
}
\keyword{data}
