% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liwcalike.R
\name{liwcalike}
\alias{liwcalike}
\alias{liwcalike.corpus}
\alias{liwcalike.character}
\title{Analyze text in a LIWC-alike fashion}
\usage{
liwcalike(x, ...)

\method{liwcalike}{corpus}(x, ...)

\method{liwcalike}{character}(
  x,
  dictionary = NULL,
  tolower = TRUE,
  verbose = TRUE,
  digits = 2,
  ...
)
}
\arguments{
\item{x}{input object, a \pkg{quanteda} \link[quanteda]{corpus} or character
vector for analysis}

\item{...}{options passed to \code{\link[quanteda]{tokens}} offering
finer-grained control over how "words" are defined}

\item{dictionary}{a \pkg{quanteda} \link[quanteda]{dictionary} object
supplied for analysis}

\item{tolower}{convert to common (lower) case before tokenizing}

\item{verbose}{if \code{TRUE} print status messages during processing}

\item{digits}{how many significant digits to print for percentage quantities}
}
\value{
a data.frame object containing the analytic results, one row per
  document supplied
}
\description{
Analyze a set of texts to produce a dataset of percentages and other
quantities describing the text, similar to the functionality supplied by the
Linguistic Inquiry and Word Count standalone software distributed at
\url{http://liwc.wpengine.com}.
}
\section{Segmentation}{
 The LIWC standalone software has many options for
  segmenting the text.  While this function does not supply segmentation
  options, you can easily achieve the same effect by converting the input
  object into a corpus (if it is not already a corpus) and using
  \code{\link[quanteda]{corpus_reshape}} or \code{\link[quanteda]{corpus_segment}}
  to split the input texts into smaller units based on user-supplied tags,
  sentence, or paragraph boundaries.
}

\examples{
liwcalike(data_char_testphrases)

# examples for comparison
txt <- c("The red-shirted lawyer gave her yellow-haired, red nose ex-boyfriend $300
            out of pity:(.")
dict <- quanteda::dictionary(list(people = c("lawyer", "boyfriend"),
                                  color_fixed = "red",
                                  color_glob = c("red*", "yellow*", "green*"),
                                  mwe = "out of"))
liwcalike(txt, dict, what = "word")
liwcalike(txt, dict, what = "fasterword")
(toks <- quanteda::tokens(txt, what = "fasterword", remove_hyphens = TRUE))
length(toks[[1]])
# LIWC says 12 words

\dontrun{# works with LIWC 2015 dictionary too
dict_liwc_2015 <-
    dictionary(file = "~/Dropbox/QUANTESS/dictionaries/LIWC/LIWC2015_English_Flat.dic",
               format = "LIWC")
dat_liwc_analysis <- liwcalike(data_corpus_inaugural, dict_liwc_2015)
dat_liwc_analysis[1:6, 1:10]
##           docname Segment      WPS   WC Sixltr    Dic function pronoun ppron    i
## 1 1789-Washington       1 66.91304 1539  24.37 240.74    52.44   12.02  5.98 3.70
## 2 1793-Washington       2 36.75000  147  25.17 236.73    53.06   12.24  6.80 6.12
## 3      1797-Adams       3 69.81081 2583  24.62 226.40    49.13    8.52  3.33 1.32
## 4  1801-Jefferson       4 47.09756 1931  20.40 241.12    49.56   12.53  6.37 1.76
## 5  1805-Jefferson       5 52.88889 2380  22.98 241.93    51.26   13.95  7.39 1.55
## 6    1809-Madison       6 60.23810 1265  24.82 239.29    52.33    9.72  4.98 2.53
}
}
